+++ src/main/java/de/christian2003/psychiatric/application/services/PatientService.java
@@ -2,13 +2,10 @@ package de.christian2003.psychiatric.application.services;
 
 import de.christian2003.psychiatric.application.repositories.CrisisInterventionAreaRepository;
 import de.christian2003.psychiatric.application.repositories.PatientRepository;
-import de.christian2003.psychiatric.application.repositories.SavableRepository;
 import de.christian2003.psychiatric.domain.people.Patient;
 import de.christian2003.psychiatric.domain.people.PersonalData;
 import de.christian2003.psychiatric.domain.rooms.CrisisInterventionArea;
-
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -62,7 +59,7 @@ public class PatientService {
     }
 
 
-    public Patient editPatient(UUID id, String firstname, String lastname, LocalDate birthday) throws NullPointerException, ServiceException {
+    public Patient editPatient(UUID id, String newFirstname, String newLastname, LocalDate newBirthday) throws NullPointerException, ServiceException {
         if (id == null) {
             throw new NullPointerException();
         }
@@ -72,17 +69,17 @@ public class PatientService {
             throw new ServiceException("Cannot edit patient, since no patient with ID \"" + id + "\" exists.");
         }
 
-        if (firstname == null) {
-            firstname = patient.getPersonalData().getFirstname();
+        if (newFirstname == null) {
+            newFirstname = patient.getPersonalData().getFirstname();
         }
-        if (lastname == null) {
-            lastname = patient.getPersonalData().getLastname();
+        if (newLastname == null) {
+            newLastname = patient.getPersonalData().getLastname();
         }
-        if (birthday == null) {
-            birthday = patient.getPersonalData().getBirthday();
+        if (newBirthday == null) {
+            newBirthday = patient.getPersonalData().getBirthday();
         }
 
-        PersonalData personalData = new PersonalData(firstname, lastname, birthday);
+        PersonalData personalData = new PersonalData(newFirstname, newLastname, newBirthday);
         patient.updatePersonalData(personalData);
         patientRepository.insertPatient(patient);
 
@@ -90,7 +87,15 @@ public class PatientService {
     }
 
 
-    public void deletePatient(Patient patient) throws NullPointerException {
+    public Patient deletePatient(UUID patientId) throws NullPointerException, ServiceException {
+        if (patientId == null) {
+            throw new NullPointerException();
+        }
+
+        Patient patient = patientRepository.getPatientById(patientId);
+        if (patient == null) {
+            throw new ServiceException("Cannot delete patient, since no patient with ID \"" + patientId + "\" exists.");
+        }
         patientRepository.deletePatient(patient);
 
         //Remove patient from assigned crisis intervention area:
@@ -100,6 +105,47 @@ public class PatientService {
                 crisisInterventionArea.removeAssignedPatient();
             }
         }
+
+        return patient;
+    }
+
+
+    public void movePatientToCrisisInterventionArea(UUID patientId, UUID crisisInterventionAreaId) throws NullPointerException, ServiceException {
+        if (patientId == null || crisisInterventionAreaId == null) {
+            throw new NullPointerException();
+        }
+
+        //Test whether patient exists:
+        Patient patient = patientRepository.getPatientById(patientId);
+        if (patient == null) {
+            throw new ServiceException("Cannot move patient, since no patient with ID \"" + patientId + "\" exists.");
+        }
+
+        //Test whether crisis intervention area exists and is free:
+        CrisisInterventionArea crisisInterventionArea = crisisInterventionAreaRepository.getCrisisInterventionAreaById(crisisInterventionAreaId);
+        if (crisisInterventionArea == null) {
+            throw new ServiceException("Cannot move patient, since no crisis intervention area with ID \"" + crisisInterventionAreaId + "\" exists.");
+        }
+        if (crisisInterventionArea.hasAssignedPatient()) {
+            if (crisisInterventionArea.getAssignedPatient().equals(patientId)) {
+                throw new ServiceException("Patient \"" + patientId + "\" is already assigned to crisis intervention area \"" + crisisInterventionAreaId + "\".");
+            }
+            else {
+                throw new ServiceException("Cannot move patient, since crisis intervention area with ID \"" + crisisInterventionAreaId + "\" already has assigned patient.");
+            }
+        }
+
+        //Remove patient from all current crisis intervention areas:
+        List<CrisisInterventionArea> crisisInterventionAreas = crisisInterventionAreaRepository.getAllCrisisInterventionAreas();
+        for (CrisisInterventionArea cia : crisisInterventionAreas) {
+            if (cia.hasAssignedPatient() && cia.getAssignedPatient().equals(patientId)) {
+                cia.removeAssignedPatient();
+            }
+        }
+
+        //Assign patient to crisis intervention area:
+        crisisInterventionArea.assignPatient(patientId);
+        crisisInterventionAreaRepository.insertCrisisInterventionArea(crisisInterventionArea);
     }
 
 }
